+++ tests/unit/orm/test_postgres_orm.py
@@ -129,6 +129,7 @@ def _test_create_schema_enum(orm, test_func, enum_name,
 # Order of parameters must match order in _create_schemas() due to dependencies
 @pytest.mark.parametrize('method_name, enum_name', [
     ('_create_schema_enum_market', 'market'),
+    ('_create_schema_enum_currency', 'currency'),
 ])
 def test__create_schemas_enums(pg_test_orm, method_name, enum_name):
     """
@@ -210,6 +211,7 @@ def _test_create_schema_table(orm, test_func, table_name,
     ('_create_schema_table_datafeed_src', 'datafeed_src'),
     ('_create_schema_table_exchange', 'exchange'),
     ('_create_schema_table_company', 'company'),
+    ('_create_schema_table_security', 'security'),
 ])
 def test__create_schemas_tables(pg_test_orm, method_name, table_name):
     """